Update complete tree from Gentoo

+++ x11-libs/libfm/libfm-9999.ebuild
@@ -1,30 +1,37 @@
-# Copyright 1999-2012 Gentoo Foundation
+# Copyright 1999-2014 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/x11-libs/libfm/libfm-9999.ebuild,v 1.32 2012/12/23 14:40:40 hwoarang Exp $
+# $Header: /var/cvsroot/gentoo-x86/x11-libs/libfm/libfm-9999.ebuild,v 1.40 2014/03/16 10:32:42 hwoarang Exp $
 
 EAPI=5
 
-EGIT_REPO_URI="git://pcmanfm.git.sourceforge.net/gitroot/pcmanfm/${PN}"
-
+EGIT_REPO_URI="https://github.com/lxde/${PN}"
+# master seems way too unstable for us to use
+EGIT_BRANCH="1.1"
 inherit autotools git-2 fdo-mime vala
 
 DESCRIPTION="A library for file management"
 HOMEPAGE="http://pcmanfm.sourceforge.net/"
 
 LICENSE="GPL-2"
-SLOT="0"
-IUSE="debug doc examples vala"
+SLOT="0/4.0.0" #copy ABI_VERSION because it seems upstream change it randomly
+IUSE="+automount debug doc examples udisks vala"
 KEYWORDS=""
 
 COMMON_DEPEND=">=dev-libs/glib-2.18:2
 	>=x11-libs/gtk+-2.16:2
 	>=lxde-base/menu-cache-0.3.2:="
 RDEPEND="${COMMON_DEPEND}
+	!lxde-base/lxshortcut
 	x11-misc/shared-mime-info
-	|| ( gnome-base/gvfs[udev,udisks] gnome-base/gvfs[udev,gdu] )"
+	automount? (
+		udisks? ( || (
+			gnome-base/gvfs[udev,udisks]
+			gnome-base/gvfs[udev,gdu]
+		) )
+		!udisks? ( gnome-base/gvfs[udev] )
+	)"
 DEPEND="${COMMON_DEPEND}
-	>=dev-lang/vala-0.14.0
-	dev-util/gtk-doc-am
+	vala? ( $(vala_depend) )
 	doc? (
 		dev-util/gtk-doc
 	)
@@ -34,25 +41,38 @@ DEPEND="${COMMON_DEPEND}
 
 DOCS=( AUTHORS TODO )
 
+REQUIRED_USE="udisks? ( automount )"
+
 src_prepare() {
 	if ! use doc; then
-		sed -ie "/^SUBDIRS_DOCS/d" Makefile.am || die "sed failed"
-		sed -ie "/^[[:space:]]*docs/d" configure.ac || die "sed failed"
+		sed -ie '/^SUBDIR.*=/s#docs##' "${S}"/Makefile.am || die "sed failed"
+		sed -ie '/^[[:space:]]*docs/d' configure.ac || die "sed failed"
 	else
 		gtkdocize --copy || die
 	fi
+	sed -i -e "s:-O0::" -e "/-DG_ENABLE_DEBUG/s: -g::" \
+		configure.ac || die "sed failed"
+
 	intltoolize --force --copy --automake || die
 	#disable unused translations. Bug #356029
 	for trans in app-chooser ask-rename exec-file file-prop preferred-apps \
 		progress;do
 		echo "data/ui/"${trans}.ui >> po/POTFILES.in
 	done
-	sed -i -e "s:-O0::" -e "/-DG_ENABLE_DEBUG/s: -g::" "${S}"/configure.ac || die
 	#Remove -Werror for automake-1.12. Bug #421101
 	sed -i "s:-Werror::" configure.ac || die
-	# Drop the symlink. We will create it later on. 439580
-	sed -i "/@LN_S@ @PACKAGE@-@FMLIBVER@/d" src/Makefile.am \
-		|| die "failed to remove the includedir symlink"
+
+	# subslot sanity check
+	local sub_slot=${SLOT#*/}
+	local libfm_major_abi=$(sed -rne '/ABI_VERSION/s:.*=::p' src/Makefile.am | tr ':' '.')
+
+	if [[ ${sub_slot} != ${libfm_major_abi} ]]; then
+		eerror "Ebuild sub-slot (${sub_slot}) does not match ABI_VERSION(${libfm_major_abi})"
+		eerror "Please update SLOT variable as follows:"
+		eerror "    SLOT=\"${SLOT%%/*}/${libfm_major_abi}\""
+		eerror
+		die "sub-slot sanity check failed"
+	fi
 
 	eautoreconf
 	rm -r autom4te.cache || die
@@ -64,9 +84,9 @@ src_configure() {
 		--sysconfdir="${EPREFIX}/etc" \
 		--disable-dependency-tracking \
 		--disable-static \
-		--disable-udisks \
 		$(use_enable examples demo) \
 		$(use_enable debug) \
+		$(use_enable udisks) \
 		$(use_enable vala actions) \
 		$(use_enable doc gtk-doc) \
 		--with-html-dir=/usr/share/doc/${PF}/html
@@ -75,9 +95,13 @@ src_configure() {
 src_install() {
 	default
 	find "${D}" -name '*.la' -exec rm -f '{}' +
-	# 439570
-	[[ -d ${D}/usr/include/${PN} ]] \
-		&& rm -r ${D}/usr/include/${PN}
+	# Remove broken symlink #439570
+	# Sometimes a directory is created instead of a symlink. No idea why...
+	# It is wrong anyway. We expect a libfm-1.0 directory and then a libfm
+	# symlink to it.
+	if [[ -h ${D}/usr/include/${PN} || -d ${D}/usr/include/${PN} ]]; then
+		rm -r "${D}"/usr/include/${PN}
+	fi
 }
 
 pkg_preinst() {
@@ -85,7 +109,7 @@ pkg_preinst() {
 	[[ -d "${ROOT}"/usr/include/${PN} ]] && \
 		rm -rf "${ROOT}"/usr/include/${PN}
 	if [[ -d "${D}"/usr/include/${PN}-1.0 ]]; then
-		cd ${D}/usr/include
+		cd "${D}"/usr/include
 		ln -s --force ${PN}-1.0 ${PN}
 	fi
 }
